Move module into backports namespace
This will make it somewhat more difficult to keep the backport version
in sync with the standard library version, but it is an important
preparation for a PyPI release.
