+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
-name = zoneinfo
-version = attr:zoneinfo._version.__version__
+name = backports.zoneinfo
+version = attr:backports.zoneinfo._version.__version__
 description = Backport of the standard library zoneinfo module
 long_description = file: README.md
 long_description_content_type = text/markdown
