+++ setup.py
@@ -4,7 +4,7 @@ import setuptools
 from setuptools import Extension
 
 c_extension = Extension(
-    "zoneinfo._czoneinfo", sources=["lib/zoneinfo_module.c"],
+    "backports.zoneinfo._czoneinfo", sources=["lib/zoneinfo_module.c"],
 )
 
 setuptools.setup(ext_modules=[c_extension])
