+++ tox.ini
@@ -20,7 +20,7 @@ setenv =
     COVERAGE_FILE={toxworkdir}/.coverage/.coverage.{envname}
     CFLAGS={env:CFLAGS:-UNDEBUG}
 commands =
-    pytest {toxinidir} {posargs: --cov=zoneinfo --cov=tests}
+    pytest {toxinidir} {posargs: --cov=backports.zoneinfo --cov=tests}
 
 # This should probably be integrated into the main testenv as an option if such
 # a thing is possible, but because it's more important that it runs on CI and
@@ -115,7 +115,7 @@ description = Run mypy on the testing example
 deps =
     mypy
 commands =
-    mypy src/zoneinfo/__init__.pyi
+    mypy src/backports/zoneinfo/__init__.pyi
     mypy tests/typing_example.py
     mypy scripts/update_test_data.py
